keyboard: cypress: Don't renew notification timeout option
The usual behaviour of BLN implementations is that notification
timeouts are renewed with each new notification event. I am not
sure if this is desired.

This patch allows the user to disable notification renewal by
echo'ing a boolean value as third parameter into notification_timeout.
The default will be standard behaviour though (renew/1).